Imrpoving summary format

+++ tasks/download_file_task.py
@@ -1,4 +1,5 @@
 import os
+from pathlib import Path
 import time
 
 import requests
@@ -10,22 +11,22 @@ class DownloadFileTask(BaseTask):
     '''
     Download a file from a given URL and save it to the specified output path.
     '''
-    def __init__(self, input_path_and_file_name:str, temp_output_path_and_file_name:str, output_path_and_file_name:str, token:str, retry_delay, max_retry, file_size=0):
+    def __init__(self, url:str, temp_output_path_and_file_name:str, output_path_and_file_name:str, token:str, retry_delay, max_retry, file_size=0):
+        self.url = url
         self.temp_output_path_and_file_name  = temp_output_path_and_file_name
-
-        if os.path.exists(self.temp_output_path_and_file_name):
-            super().__init__(f'Resume Download Of: \"{input_path_and_file_name}\" to: \"{output_path_and_file_name}\"', input_path_and_file_name, output_path_and_file_name)
-        else:
-            super().__init__(f'Download Of: \"{input_path_and_file_name}\" to: \"{output_path_and_file_name}\"', input_path_and_file_name, output_path_and_file_name)
-
         self.token = token
         self.file_size = file_size
         self.retry_delay = retry_delay
         self.max_retry = max_retry
 
+        if os.path.exists(self.temp_output_path_and_file_name):
+            super().__init__(f'Resume Download File From: \"{url}\" to: \"{Path(temp_output_path_and_file_name).name}\" when complete move to \"{Path(output_path_and_file_name).name}\"')
+        else:
+            super().__init__(f'Download File: \"{url}\" to: \"{Path(temp_output_path_and_file_name).name}\" when complete move to \"{Path(output_path_and_file_name).name}\"')
+
 
     def run(self):
-        self.logger.debug('Downloading: %s to %s, Resuming? %s', self.input_path_and_file_name, self.output_path_and_file_name, os.path.exists(self.temp_output_path_and_file_name))
+        self.logger.debug('Downloading: %s to %s, Resuming? %s', self.url, self.output_path_and_file_name, os.path.exists(self.temp_output_path_and_file_name))
 
         for r in range(self.max_retry):
             try:
@@ -42,18 +43,18 @@ class DownloadFileTask(BaseTask):
                     color = 'YELLOW'
 
                 # Download the file with progress bar
-                with requests.get(self.input_path_and_file_name, headers={ 'Authorization': f'Bearer {self.token}', **resume_header }, stream=True, timeout=2000, allow_redirects=True) as response:
+                with requests.get(self.url, headers={ 'Authorization': f'Bearer {self.token}', **resume_header }, stream=True, timeout=2000, allow_redirects=True) as response:
 
                     if response.status_code == 401:
-                        self.logger.debug("Unauthorized for url (Model Removed?): %s, Reason: %s", self.input_path_and_file_name, response.reason)
+                        self.logger.debug("Unauthorized for url (Model Removed?): %s, Reason: %s", self.url, response.reason)
                         return False
 
                     if response.status_code == 404:
-                        self.logger.debug("File not found (Model Removed?): %s, Reason: %s", self.input_path_and_file_name, response.reason)
+                        self.logger.debug("File not found (Model Removed?): %s, Reason: %s", self.url, response.reason)
                         return False
 
                     if response.status_code == 416:
-                        self.logger.debug("Could not resume download, resume was: (range:%s returned_length:%s) %s, Reason: %s", resume_header['Range'], int(response.headers.get('Content-Length', 0)), self.input_path_and_file_name, response.reason)
+                        self.logger.debug("Could not resume download, resume was: (range:%s returned_length:%s) %s, Reason: %s", resume_header['Range'], int(response.headers.get('Content-Length', 0)), self.url, response.reason)
                         return False
 
                     response.raise_for_status()
@@ -72,7 +73,12 @@ class DownloadFileTask(BaseTask):
 
                 # Rename the temp file to the final output path
                 self.logger.debug("Download complete, renaming %s -> %s", self.temp_output_path_and_file_name, self.output_path_and_file_name)
-                os.rename(self.temp_output_path_and_file_name, self.output_path_and_file_name)
+
+                if os.path.exists(self.temp_output_path_and_file_name):
+                    os.rename(self.temp_output_path_and_file_name, self.output_path_and_file_name)
+                else:
+                    self.logger.error("Downloaded file not found? %s", self.temp_output_path_and_file_name)
+                    return False
 
                 # Break out of retry loop if download was successful
                 return True
@@ -85,7 +91,7 @@ class DownloadFileTask(BaseTask):
                 self.logger.error("Abnormal Error Occured downloading: %s", self.output_path_and_file_name, type(e), e, stack_info=True, exc_info=True)
                 return False
 
-        self.logger.error("Failed to download file: %s, hit max retries.", self.input_path_and_file_name)
+        self.logger.error("Failed to download file: %s, hit max retries.", self.url)
         return False
